requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
gunicorn==21.2.0
idna==3.11
packaging==26.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
